fix(pathfinding): route around every kind of meeting area
The pathfinder had its own definition of a meeting area - Jitsi and Livekit
only - so avatars walked straight through a Teams or Google Meet meeting while
they routed around a Jitsi one. Reuse the shared rule.

Moved the rule out of the area editor into Rules/, now that the runtime reads
it too, and share the extension module lookup both sides were duplicating.
