+++ play/src/front/Components/MapEditor/AreaEditor/AreaPropertiesEditor.svelte
@@ -32,7 +32,6 @@
     import RightsPropertyEditor from "../PropertyEditor/RightsPropertyEditor.svelte";
     import { IconChevronDown, IconChevronRight, IconInfoCircle } from "../../Icons";
     import { extensionModuleStore } from "../../../Stores/GameSceneStore";
-    import type { ExtensionModule, ExtensionModuleAreaProperty } from "../../../ExternalModule/ExtensionModule";
     import MatrixRoomPropertyEditor from "../PropertyEditor/MatrixRoomPropertyEditor.svelte";
     import TooltipPropertyButton from "../PropertyEditor/TooltipPropertyButton.svelte";
     import LivekitRoomPropertyEditor from "../PropertyEditor/LivekitRoomPropertyEditor.svelte";
@@ -44,7 +43,7 @@
     import { gameManager } from "../../../Phaser/Game/GameManager";
     import MaxUsersInAreaPropertyEditor from "../PropertyEditor/MaxUsersInAreaPropertyEditor.svelte";
     import LockableAreaPropertyEditor from "../PropertyEditor/LockableAreaPropertyEditor.svelte";
-    import { hasMeetingProperty } from "./meetingProperties";
+    import { getAreaMapEditors, hasMeetingProperty } from "../../../Rules/MeetingRules";
 
     let properties: AreaDataProperties = $state([]);
     let areaName = $state("");
@@ -550,16 +549,7 @@
         showDescriptionField = !showDescriptionField;
     }
 
-    let extensionModulesAreaMapEditor = $extensionModuleStore.reduce(
-        (acc: { [key: string]: ExtensionModuleAreaProperty }[], module: ExtensionModule) => {
-            const areaProperty = module.areaMapEditor?.();
-            if (areaProperty != undefined) {
-                acc.push(areaProperty);
-            }
-            return acc;
-        },
-        [],
-    );
+    let extensionModulesAreaMapEditor = getAreaMapEditors($extensionModuleStore);
 
     let hasMeeting = $derived(hasMeetingProperty(properties, extensionModulesAreaMapEditor));
 </script>
